Compute shown entity attributes in one place
